Feat: add product service in api/products route

+++ prisma/migrations/20230123194653_up/migration.sql
@@ -23,7 +23,7 @@ CREATE TABLE "Product" (
     "image" TEXT NOT NULL,
     "price" INTEGER NOT NULL,
     "brand" TEXT NOT NULL,
-    "rating" INTEGER NOT NULL,
+    "rating" REAL NOT NULL,
     "numReviews" INTEGER NOT NULL,
     "countInStock" INTEGER NOT NULL,
     "description" TEXT NOT NULL,
